Correct cyclic references in SpringBoot. Update parent-web to 7.0.0, selfautorestdoc to 0.2.2, setupdb to 2.0.1, corrects warnings (deprectated, and false deprectated for Hibernate) #37
