+++ pom.xml
@@ -14,7 +14,7 @@
 	<parent>
 		<groupId>tv.hd3g.commons</groupId>
 		<artifactId>parent-web</artifactId>
-		<version>4.0.0</version>
+		<version>7.0.0</version>
 		<relativePath />
 	</parent>
 
@@ -72,7 +72,7 @@
 		<dependency>
 			<groupId>tv.hd3g</groupId>
 			<artifactId>selfautorestdoc</artifactId>
-			<version>0.0.5</version>
+			<version>0.2.2</version>
 		</dependency>
 		<dependency>
 			<groupId>com.h2database</groupId>
@@ -96,7 +96,7 @@
 			<plugin>
 				<groupId>tv.hd3g.mvnplugin</groupId>
 				<artifactId>setupdb</artifactId>
-				<version>0.0.2</version>
+				<version>2.0.1</version>
 			</plugin>
 		</plugins>
 	</build>
