+++ src/test/java/tv/hd3g/authkit/mod/controller/RestControllerMeTest.java
@@ -82,7 +82,7 @@ class RestControllerMeTest {
 	private static final ResultMatcher statusOk = status().isOk();
 	private static final ResultMatcher statusBadRequest = status().isBadRequest();
 	private static final ResultMatcher contentTypeJsonUtf8 = content().contentType(APPLICATION_JSON_VALUE);
-	private static final ResultMatcher statusOkUtf8 = ResultMatcher.matchAll(statusOk, contentTypeJsonUtf8);
+	private static final ResultMatcher[] statusOkUtf8 = new ResultMatcher[] { statusOk, contentTypeJsonUtf8 };
 
 	@Autowired
 	private MockMvc mvc;
@@ -151,7 +151,7 @@ class RestControllerMeTest {
 	void prepareTOTP() throws Exception {
 		mvc.perform(get(baseMapping + "/" + "set2auth")
 		        .headers(baseHeaders))
-		        .andExpect(statusOkUtf8)
+		        .andExpectAll(statusOkUtf8)
 		        .andExpect(jsonPath("$.secret").isString())
 		        .andExpect(jsonPath("$.totpURI").isString())
 		        .andExpect(jsonPath("$.qrcode").isString())
@@ -219,7 +219,7 @@ class RestControllerMeTest {
 
 		mvc.perform(get(baseMapping + "/" + "has2auth")
 		        .headers(baseHeaders))
-		        .andExpect(statusOkUtf8)
+		        .andExpectAll(statusOkUtf8)
 		        .andExpect(jsonPath("$.twoAuthEnabled").isBoolean())
 		        .andExpect(jsonPath("$.twoAuthEnabled").value(true));
 	}
@@ -228,7 +228,7 @@ class RestControllerMeTest {
 	void hasATOTP_no() throws Exception {
 		mvc.perform(get(baseMapping + "/" + "has2auth")
 		        .headers(baseHeaders))
-		        .andExpect(statusOkUtf8)
+		        .andExpectAll(statusOkUtf8)
 		        .andExpect(jsonPath("$.twoAuthEnabled").isBoolean())
 		        .andExpect(jsonPath("$.twoAuthEnabled").value(false));
 	}
@@ -255,7 +255,7 @@ class RestControllerMeTest {
 	void isExternalAuth_no() throws Exception {
 		mvc.perform(get(baseMapping + "/" + "is-external-auth")
 		        .headers(baseHeaders))
-		        .andExpect(statusOkUtf8)
+		        .andExpectAll(statusOkUtf8)
 		        .andExpect(jsonPath("$.externalAuthEnabled").isBoolean())
 		        .andExpect(jsonPath("$.externalAuthEnabled").value(false));
 	}
@@ -269,7 +269,7 @@ class RestControllerMeTest {
 
 		mvc.perform(get(baseMapping + "/" + "is-external-auth")
 		        .headers(baseHeaders))
-		        .andExpect(statusOkUtf8)
+		        .andExpectAll(statusOkUtf8)
 		        .andExpect(jsonPath("$.externalAuthEnabled").isBoolean())
 		        .andExpect(jsonPath("$.externalAuthEnabled").value(true))
 		        .andExpect(jsonPath("$.domain").isString())
@@ -310,7 +310,7 @@ class RestControllerMeTest {
 			        .andExpect(jsonPath("$.phone").value(expected.getPhone()))
 			        .andExpect(jsonPath("$.postalcode").value(expected.getPostalcode()))
 			        .andExpect(jsonPath("$.userUUID").value(expected.getUserUUID()))
-			        .andExpect(statusOkUtf8);
+			        .andExpectAll(statusOkUtf8);
 		}
 
 		@Test
